[Feature] 정기 소득(RecurringIncome) 엔티티 추가 및 매핑 (#23)
* feat: 정기 수입 테이블 추가 및 금액 데이터타입 통일

* feat: 명세 수정 및 데이터 타입변경

+++ backend/src/main/java/com/dday/domain/welfare/dto/response/WelfareReviewItemResponse.java
@@ -8,7 +8,6 @@ import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
 
-import java.math.BigDecimal;
 import java.time.LocalDateTime;
 import java.util.Arrays;
 import java.util.List;
@@ -35,7 +34,7 @@ public class WelfareReviewItemResponse {
     // ── 문제일 수 있는 현재 값 ──
     private String category;
     private SupportType supportType;
-    private BigDecimal supportAmount;
+    private Long supportAmount;
     private SupportAmountType supportAmountType;
     private Integer supportDurationMonths;
     private String targetDescription;
